requirements.txt
ascii-magic
attrs
autocorrect
beautifulsoup4
black
boto3
bs4
colorama
coloredlogs
#contextualSpellCheck
coverage
cryptography
envelope==2.0.0
eth-brownie
eth-utils
Faker
fake-useragent
flake8
ipfshttpclient
irc
language_tool_python
latex2mathml
lxml
git+https://github.com/strikles/markdown-pp.git@master
markovify
git+https://github.com/shawnanastasio/python-matrix-bot-api.git@master
mistune==2.0.2
nltk
numpy
pandas
phonenumbers
Pillow
py-solc-x
pydbg
pyaxmlparser
pygments
pylatexenc
pyspellchecker
pystache
pytest
python-dotenv
python-gnupg
python-magic-bin; platform_system == 'Windows'
pywhatkit
pywsd==1.0.4
PyYAML
qrcode
rdoclient
regex
requests
samila
scipy
shadow-useragent
spacy
stdiomask
syllapy
sympy
text_unidecode
tldextract
torrequest
tqdm
translators
tsp_solver2
urllib3
utils
validators
vyper
watchdog
web3
wn
